requirements.txt

aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.2.0
packaging==24.2
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-slugify==8.0.4
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.19.0
zipp==3.21.0
